- aplica correções do pull request;
- remove arquivos desnecessários;
- corrige configs do banco de dados;
- simplifica algumas configs de envs;

+++ src/main/java/br/com/alfac/videostudio/core/domain/service/LogService.java
@@ -3,5 +3,5 @@ package br.com.alfac.videostudio.core.domain.service;
 import java.time.LocalDateTime;
 
 public interface LogService {
-    void saveLog(Integer usuarioId, String resource, LocalDateTime datetime);
+    void saveLog(Long usuarioId, String resource, LocalDateTime datetime);
 }
